requirements.txt
adb-shell==0.4.3
black==22.8.0
cffi==1.15.1
click==8.0.4
colorama==0.4.5
cryptography==40.0.2
dataclasses==0.8
importlib-metadata==4.8.3
libusb1==3.0.0
MouseInfo==0.1.3
mypy-extensions==1.0.0
pathspec==0.9.0
Pillow==8.4.0
platformdirs==2.4.0
pure-python-adb==0.3.0.dev0
pyasn1==0.4.6
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytesseract==0.3.8
pytweening==1.0.4
rsa==4.9
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
zipp==3.6.0